Implement a `hyperclick` provider for jumping to a definition via mouse
The downside is that we can’t defer activation until the first command 
invocation anymore, but I’m OK with that. If we provide go-to-reference 
via the keyboard, we should provide it via mouse, too; and it makes a 
lot of sense to implement it here rather than require the user to have 
some wrapper package.